claude-code-os-implementation/03-ai-growth-engine/organic-marketing-package/packages/content-engine/requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
cryptography>=41.0.0
sqlalchemy>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
flake8>=6.1.0
python-amazon-sp-api>=1.0.0
requests>=2.31.0
APScheduler>=3.10.0
sendgrid>=6.11.0
jinja2>=3.1.0
celery[redis]>=5.3.0
redis>=5.0.0
